Add "shift" to clkkdiv_n_2_4_8
The fastest clock can now be shifted right so that clkpn is mapped to
clkp2n and clkp4n to clkp8n . Original clkp8n signal is "dropped" out
from right.

TODO: Shift could be two-directional, and the solution could be more
elegant, but did not have time to polish.